Move public-API tests to integration tests, use subtle::ConstantTimeEq
