+++ Cargo.toml
@@ -87,6 +87,7 @@ sha3 = "0.10.8"
 spmc = "=0.3.0"
 strum = "0.27.2"
 strum_macros = "0.26.1"
+subtle = { version = "2.5", default-features = false }
 thiserror = { version = "2", default-features = false }
 tikv-jemalloc-ctl = { version = "0.5" }
 tikv-jemallocator = { version = "0.5" }
