+++ crates/mina-base58/Cargo.toml
@@ -9,6 +9,7 @@ repository.workspace = true
 [dependencies]
 bs58.workspace = true
 sha2.workspace = true
+subtle = { workspace = true }
 thiserror = { workspace = true }
 
 [dev-dependencies]
